remove dependency of version.h on .git/* to avoid errors
the wildcard function in GNU make includes dangling symlinks; if any
exist under the .git directory, they would get added as dependencies,
causing make to exit with an error due to lacking a rule to build the
missing file.

as far as I can tell, git operations which should force version.h to
be rebuilt must all touch the mtime of the top-level .git directory.